vocab-quiz-functions/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-functions==1.19.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
openai==1.30.5
pandas==2.2.2
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
